jaspQmlSource should not encode Html characters
Fixes https://github.com/jasp-stats/jasp-test-release/issues/1763

The jaspQmlSource is derived from JaspJson, that has functions to convert R objects to Json (RVectorEntry_to_JsonValue & RMatrixColumnEntry_to_JsonValue). These functions encode per default Html characters (because the result is send to the Web Browser), but this is not wanted for jaspQmlSource objects. To solve this problem, an extra class property is added _encodeHtml which is set to true per default, but set to false for jaspQmlSource.
Unfortunately these functions that convert R objects to Json value are static, because these functions are directly used in other classes that are not derived from jaspJson. So all these static functions have been moved to jaspObject so that all other objects have access directly to them.